Fix warnings printed by the Microsoft compiler at -W3
 src\cbor.h(268): warning C4244: '=': conversion from 'const uint16_t' to 'uint8_t', possible loss of data
 src\cbor.h(422): warning C4244: 'initializing': conversion from 'uint64_t' to 'uint32_t', possible loss of data
 src\cborencoder.c(134): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)
 src\cbortojson.c(410): warning C4244: '=': conversion from 'uint64_t' to 'double', possible loss of data

+++ src/cbor.h
@@ -265,7 +265,7 @@ CBOR_INLINE_API bool cbor_value_is_simple_type(const CborValue *value)
 CBOR_INLINE_API CborError cbor_value_get_simple_type(const CborValue *value, uint8_t *result)
 {
     assert(cbor_value_is_simple_type(value));
-    *result = value->extra;
+    *result = (uint8_t)value->extra;
     return CborNoError;
 }
 
@@ -419,7 +419,7 @@ CBOR_INLINE_API CborError cbor_value_get_float(const CborValue *value, float *re
 {
     assert(value->type == CborFloatType);
     assert(value->flags & CborIteratorFlag_IntegerValueTooLarge);
-    uint32_t data = _cbor_value_decode_int64_internal(value);
+    uint32_t data = (uint32_t)_cbor_value_decode_int64_internal(value);
     memcpy(result, &data, sizeof(*result));
     return CborNoError;
 }
